Move IItemWithTier static methods to implementation interface
