+++ src/main/java/de/teamlapen/vampirism/items/IItemWithTierNBTImpl.java
@@ -1,8 +1,10 @@
 package de.teamlapen.vampirism.items;
 
 import de.teamlapen.lib.lib.util.UtilLib;
+import de.teamlapen.vampirism.VampirismMod;
 import de.teamlapen.vampirism.api.items.IItemWithTier;
 import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -15,6 +17,26 @@ import java.util.List;
  */
 public interface IItemWithTierNBTImpl extends IItemWithTier {
 
+    static TIER getTierStatic(@Nonnull ItemStack stack) {
+        NBTTagCompound tag = UtilLib.checkNBT(stack);
+        if (tag.hasKey("tier")) {
+            try {
+                return TIER.valueOf(tag.getString("tier"));
+            } catch (IllegalArgumentException e) {
+                VampirismMod.log.e(stack.getItem().getRegistryName().toString(), e, "Unknown item tier %s", tag.getString("tier"));
+            }
+
+        }
+        return TIER.NORMAL;
+    }
+
+    @Nonnull
+    static ItemStack setTierStatic(@Nonnull ItemStack stack, TIER tier) {
+        NBTTagCompound tag = UtilLib.checkNBT(stack);
+        tag.setString("tier", tier.name());
+        return stack;
+    }
+
     @SideOnly(Side.CLIENT)
     default void addTierInformation(ItemStack stack, List<String> tooltip) {
         TIER t = getTier(stack);
@@ -25,12 +47,12 @@ public interface IItemWithTierNBTImpl extends IItemWithTier {
 
     @Override
     default TIER getTier(@Nonnull ItemStack stack) {
-        return IItemWithTier.getTierStatic(stack);
+        return IItemWithTierNBTImpl.getTierStatic(stack);
     }
 
     @Nonnull
     @Override
     default ItemStack setTier(@Nonnull ItemStack stack, TIER tier) {
-        return IItemWithTier.setTierStatic(stack, tier);
+        return IItemWithTierNBTImpl.setTierStatic(stack, tier);
     }
 }
